Move classes into existing structure.
